Update opcode constants from i32 to i8 for memory optimization and consistency

+++ knight-age-server/src/network/opcode.rs
@@ -1,221 +1,221 @@
-const MINI_GAME: i32 = -91;
+const MINI_GAME: i8 = -91;
 
-const NOKIA_PURCHASE_MESSAGE: i32 = -76;
+const NOKIA_PURCHASE_MESSAGE: i8 = -76;
 
-const GOOGLE_PURCHASE_MESSAGE: i32 = -75;
+const GOOGLE_PURCHASE_MESSAGE: i8 = -75;
 
-const UPDATE_DATA: i32 = -57;
+const UPDATE_DATA: i8 = -57;
 
-const NAP_DIAMOND: i32 = -56;
+const NAP_DIAMOND: i8 = -56;
 
-const CHECK_UPDATE_DATA: i32 = -54;
+const CHECK_UPDATE_DATA: i8 = -54;
 
-const NAP_TIEN: i32 = -53;
+const NAP_TIEN: i8 = -53;
 
-const LOAD_IMAGE_DATA_PART_CHAR: i32 = -52;
+const LOAD_IMAGE_DATA_PART_CHAR: i8 = -52;
 
-const LOAD_IMAGE: i32 = -51;
+const LOAD_IMAGE: i8 = -51;
 
-const NPC_BIG: i32 = -50;
+const NPC_BIG: i8 = -50;
 
-const LOAD_IMAGE_DATA_AUTO_EFF: i32 = -49;
+const LOAD_IMAGE_DATA_AUTO_EFF: i8 = -49;
 
-const NEW_NPC_INFO: i32 = -44;
+const NEW_NPC_INFO: i8 = -44;
 
-const DIALOG_MORE_OPTION_SERVER: i32 = -31;
+const DIALOG_MORE_OPTION_SERVER: i8 = -31;
 
-const DIALOG_SERVER: i32 = -32;
+const DIALOG_SERVER: i8 = -32;
 
-const DYNAMIC_MENU: i32 = -30;
+const DYNAMIC_MENU: i8 = -30;
 
-const MONSTER_CAPCHAR: i32 = -28;
+const MONSTER_CAPCHAR: i8 = -28;
 
-const LOGIN: i32 = 1;
+const LOGIN: i8 = 1;
 
-const LOGIN_FAIL: i32 = 2;
+const LOGIN_FAIL: i8 = 2;
 
-const MAIN_CHAR_INFO: i32 = 3;
+const MAIN_CHAR_INFO: i8 = 3;
 
-const OBJECT_MOVE: i32 = 4;
+const OBJECT_MOVE: i8 = 4;
 
-const CHAR_INFO: i32 = 5;
+const CHAR_INFO: i8 = 5;
 
-const FIRE_PK: i32 = 6;
+const FIRE_PK: i8 = 6;
 
-const MONSTER_INFO: i32 = 7;
+const MONSTER_INFO: i8 = 7;
 
-const PLAYER_EXIT: i32 = 8;
+const PLAYER_EXIT: i8 = 8;
 
-const FIRE_MONSTER: i32 = 9;
+const FIRE_MONSTER: i8 = 9;
 
-const MONSTER_FIRE: i32 = 10;
+const MONSTER_FIRE: i8 = 10;
 
-const USE_ITEM: i32 = 11;
+const USE_ITEM: i8 = 11;
 
-const CHANGE_MAP: i32 = 12;
+const CHANGE_MAP: i8 = 12;
 
-const LIST_CHAR: i32 = 13;
+const LIST_CHAR: i8 = 13;
 
-const SELECT_CHAR: i32 = 13;
+const SELECT_CHAR: i8 = 13;
 
-const CREATE_CHAR: i32 = 14;
+const CREATE_CHAR: i8 = 14;
 
-const CHAR_WEARING: i32 = 15;
+const CHAR_WEARING: i8 = 15;
 
-const CHAR_INVENTORY: i32 = 16;
+const CHAR_INVENTORY: i8 = 16;
 
-const DIE_MONSTER: i32 = 17;
+const DIE_MONSTER: i8 = 17;
 
-const DEL_ITEM: i32 = 18;
+const DEL_ITEM: i8 = 18;
 
-const ITEM_DROP: i32 = 19;
+const ITEM_DROP: i8 = 19;
 
-const GET_ITEM_MAP: i32 = 20;
+const GET_ITEM_MAP: i8 = 20;
 
-const ITEM_MORE_INFO: i32 = 21;
+const ITEM_MORE_INFO: i8 = 21;
 
-const ADD_BASE_SKILL_POINT: i32 = 22;
+const ADD_BASE_SKILL_POINT: i8 = 22;
 
-const NPC_INFO: i32 = 23;
+const NPC_INFO: i8 = 23;
 
-const BUY_ITEM: i32 = 24;
+const BUY_ITEM: i8 = 24;
 
-const ITEM_TEMPLATE: i32 = 25;
+const ITEM_TEMPLATE: i8 = 25;
 
-const CATALORY_MONSTER: i32 = 26;
+const CATALORY_MONSTER: i8 = 26;
 
-const CHAT_POPUP: i32 = 27;
+const CHAT_POPUP: i8 = 27;
 
-const GET_ITEM_TEM: i32 = 28;
+const GET_ITEM_TEM: i8 = 28;
 
-const LIST_SKILL: i32 = 29;
+const LIST_SKILL: i8 = 29;
 
-const SET_EXP: i32 = 30;
+const SET_EXP: i8 = 30;
 
-const GO_HOME: i32 = 31;
+const GO_HOME: i8 = 31;
 
-const USE_POTION: i32 = 32;
+const USE_POTION: i8 = 32;
 
-const LEVEL_UP: i32 = 33;
+const LEVEL_UP: i8 = 33;
 
-const CHAT_TAB: i32 = 34;
+const CHAT_TAB: i8 = 34;
 
-const FRIEND: i32 = 35;
+const FRIEND: i8 = 35;
 
-const BUY_SELL: i32 = 36;
+const BUY_SELL: i8 = 36;
 
-const INFO_FROM_SERVER: i32 = 37;
+const INFO_FROM_SERVER: i8 = 37;
 
-const REGISTER: i32 = 39;
+const REGISTER: i8 = 39;
 
-const BUFF: i32 = 40;
+const BUFF: i8 = 40;
 
-const DIE_PLAYER: i32 = 41;
+const DIE_PLAYER: i8 = 41;
 
-const PK: i32 = 42;
+const PK: i8 = 42;
 
-const UPDATE_PET_CONTAINER: i32 = 44;
+const UPDATE_PET_CONTAINER: i8 = 44;
 
-const PET_EAT: i32 = 45;
+const PET_EAT: i8 = 45;
 
-const PARTY: i32 = 48;
+const PARTY: i8 = 48;
 
-const OTHER_PLAYER_INFO: i32 = 49;
+const OTHER_PLAYER_INFO: i8 = 49;
 
-const EFF_PLUS_TIME: i32 = 50;
+const EFF_PLUS_TIME: i8 = 50;
 
-const CHANGE_AREA: i32 = 51;
+const CHANGE_AREA: i8 = 51;
 
-const QUEST: i32 = 52;
+const QUEST: i8 = 52;
 
-const INFO_EASY_SERVER: i32 = 53;
+const INFO_EASY_SERVER: i8 = 53;
 
-const UPDATE_STATUS_AREA: i32 = 54;
+const UPDATE_STATUS_AREA: i8 = 54;
 
-const SAVE_RMS: i32 = 55;
+const SAVE_RMS: i8 = 55;
 
-const LIST_SERVER: i32 = 56;
+const LIST_SERVER: i8 = 56;
 
-const LIST_PLAYER_PK: i32 = 57;
+const LIST_PLAYER_PK: i8 = 57;
 
-const PLAYER_SUCKHOE: i32 = 59;
+const PLAYER_SUCKHOE: i8 = 59;
 
-const CHAT_NPC: i32 = 60;
+const CHAT_NPC: i8 = 60;
 
-const NAME_SERVER: i32 = 61;
+const NAME_SERVER: i8 = 61;
 
-const X2_XP: i32 = 62;
+const X2_XP: i8 = 62;
 
-const DELETE_RMS: i32 = 63;
+const DELETE_RMS: i8 = 63;
 
-const HELP_FROM_SERVER: i32 = 64;
+const HELP_FROM_SERVER: i8 = 64;
 
-const UPDATE_CHAR_CHEST: i32 = 65;
+const UPDATE_CHAR_CHEST: i8 = 65;
 
-const REBUILD_ITEM: i32 = 67;
+const REBUILD_ITEM: i8 = 67;
 
-const THACH_DAU: i32 = 68;
+const THACH_DAU: i8 = 68;
 
-const CLAN: i32 = 69;
+const CLAN: i8 = 69;
 
-const UPDATE_HP_NPC: i32 = 70;
+const UPDATE_HP_NPC: i8 = 70;
 
-const CHAT_WORLD: i32 = 71;
+const CHAT_WORLD: i8 = 71;
 
-const REPLACE_PLUS_ITEM: i32 = 73;
+const REPLACE_PLUS_ITEM: i8 = 73;
 
-const SHOW_NUM_EFF: i32 = 74;
+const SHOW_NUM_EFF: i8 = 74;
 
-const EFF_SERVER: i32 = 75;
+const EFF_SERVER: i8 = 75;
 
-const EFF_WEATHER: i32 = 76;
+const EFF_WEATHER: i8 = 76;
 
-const REBUILD_WING: i32 = 77;
+const REBUILD_WING: i8 = 77;
 
-const OPEN_BOX: i32 = 78;
+const OPEN_BOX: i8 = 78;
 
-const PET_ATTACK: i32 = 84;
+const PET_ATTACK: i8 = 84;
 
-const MONSTER_DETONATE: i32 = 85;
+const MONSTER_DETONATE: i8 = 85;
 
-const MONSTER_SKILL_INFO: i32 = 86;
+const MONSTER_SKILL_INFO: i8 = 86;
 
-const PET_GAIN_XP: i32 = 87;
+const PET_GAIN_XP: i8 = 87;
 
-const REMOVE_ACTOR: i32 = 90;
+const REMOVE_ACTOR: i8 = 90;
 
-const IN_APP_PURCHASE: i32 = -93;
+const IN_APP_PURCHASE: i8 = -93;
 
-const USE_ITEM_ARENA: i32 = -92;
+const USE_ITEM_ARENA: i8 = -92;
 
-const STATUS_ARENA: i32 = -94;
+const STATUS_ARENA: i8 = -94;
 
-const MARKKILLER: i32 = -95;
+const MARKKILLER: i8 = -95;
 
-const SERVER_ADD_NPC: i32 = -96;
+const SERVER_ADD_NPC: i8 = -96;
 
-const USE_MOUNT: i32 = -97;
+const USE_MOUNT: i8 = -97;
 
-const USE_SHIP: i32 = -98;
+const USE_SHIP: i8 = -98;
 
-const LAST_LOGIN: i32 = -99;
+const LAST_LOGIN: i8 = -99;
 
-const KHAM_NGOC: i32 = -100;
+const KHAM_NGOC: i8 = -100;
 
-const COMPETITION: i32 = -101;
+const COMPETITION: i8 = -101;
 
-const MUA_BAN: i32 = -102;
+const MUA_BAN: i8 = -102;
 
-const INFO_MI_NUONG: i32 = -103;
+const INFO_MI_NUONG: i8 = -103;
 
-const UPDATE_INFO_CLAN_HOLD: i32 = -104;
+const UPDATE_INFO_CLAN_HOLD: i8 = -104;
 
-const HOP_RAC: i32 = -105;
+const HOP_RAC: i8 = -105;
 
-const GET_MATERITAL_TEMPLATE: i32 = -106;
+const GET_MATERITAL_TEMPLATE: i8 = -106;
 
-const USE_MATERIAL: i32 = -107;
+const USE_MATERIAL: i8 = -107;
 
-const FILL_REC_UPDATE_TIME: i32 = -108;
+const FILL_REC_UPDATE_TIME: i8 = -108;
 
-const GET_NAP_STORE_APPLE: i32 = -109;
+const GET_NAP_STORE_APPLE: i8 = -109;
